rework webpack config files
