--- web-app/build-utils/webpack.common.js
@@ -1,37 +1,0 @@
-const commonPaths = require('./common-paths');
-
-const HtmlWebpackPlugin = require('html-webpack-plugin');
-const MomentLocalesPlugin = require('moment-locales-webpack-plugin');
-const MomentTimezoneDataPlugin = require('moment-timezone-data-webpack-plugin');
-
-const config = {
-  output: {
-    path: commonPaths.outputPath,
-    publicPath: '/'
-  },
-  resolve: {
-    extensions: ['.js', '.jsx', '.ts', '.tsx'],
-  },
-  module: {
-    rules: [
-      {
-        test: /\.[tj]sx?$/,
-        exclude: /node_modules/,
-        loader: 'babel-loader',
-      },
-    ]
-  },
-  plugins: [
-    new HtmlWebpackPlugin({
-      template: `public/index.html`,
-    }),
-    new MomentLocalesPlugin(),
-    new MomentTimezoneDataPlugin({
-      matchZones: ['Etc/UTC'],
-      startYear: 2020,
-      endYear: 2025,
-    }),
-  ]
-};
-
-module.exports = config;
