--- web-app/build-utils/webpack.dev.js
@@ -1,69 +1,0 @@
-const commonPaths = require('./common-paths');
-const webpack = require('webpack');
-
-const port = process.env.PORT || 3000;
-
-const config = {
-  mode: 'development',
-  entry: {
-    app: `${commonPaths.appEntry}/index.tsx`,
-  },
-  output: {
-    filename: '[name].[fullhash].js',
-  },
-  devtool: 'eval',
-  module: {
-    rules: [
-      {
-        test: /\.s[ac]ss$/,
-        use: [
-          {
-            loader: 'style-loader',
-            options: {
-              esModule: true,
-              modules: {
-                namedExport: true,
-              },
-            },
-          },
-          {
-            loader: 'css-loader',
-            options: {
-              esModule: true,
-              modules: {
-                namedExport: true,
-              },
-            },
-          },
-          {
-            loader: "sass-loader",
-            options: {
-              // Prefer `dart-sass`
-              implementation: require("sass"),
-            },
-          },
-        ],
-      },
-    ],
-  },
-  plugins: [
-    new webpack.HotModuleReplacementPlugin(),
-    new webpack.ProvidePlugin({
-      'PropTypes': 'prop-types',
-    }),
-  ],
-  resolve: {
-    alias: {
-      'react-dom': '@hot-loader/react-dom'
-    }
-  },
-  devServer: {
-    host: 'localhost',
-    port: port,
-    historyApiFallback: true,
-    hot: true,
-    open: true,
-  },
-};
-
-module.exports = config;
