--- web-app/build-utils/webpack.prod.js
@@ -1,69 +1,0 @@
-const commonPaths = require('./common-paths');
-
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-
-const config = {
-  mode: 'production',
-  entry: {
-    app: [`${commonPaths.appEntry}/index.tsx`],
-  },
-  output: {
-    filename: 'static/[name].[fullhash].js',
-  },
-  module: {
-    rules: [
-      {
-        test: /\.s[ac]ss$/,
-        use: [
-          {
-            loader: MiniCssExtractPlugin.loader,
-            options: {
-              esModule: true,
-              modules: {
-                namedExport: true,
-              },
-            },
-          },
-          {
-            loader: 'css-loader',
-            options: {
-              esModule: true,
-              modules: {
-                namedExport: true,
-              },
-            }
-          },
-          {
-            loader: 'postcss-loader',
-            options: {
-              postcssOptions: {
-                plugins: [
-                  [
-                    "postcss-preset-env",
-                    {
-                      // Options
-                    },
-                  ],
-                ],
-              },
-            },
-          },
-          {
-            loader: "sass-loader",
-            options: {
-              // Prefer `dart-sass`
-              implementation: require("sass"),
-            },
-          },
-        ],
-      },
-    ],
-  },
-  plugins: [
-    new MiniCssExtractPlugin({
-      filename: 'styles/[name].[fullhash].css',
-    }),
-  ],
-};
-
-module.exports = config;
