+++ web-app/webpack.config.js
@@ -1,5 +1,5 @@
 const chalk = require('chalk');
-const commonConfig = require('./build-utils/webpack.common.js');
+const commonConfig = require('./webpack/webpack.common.js');
 
 const webpack = require('webpack');
 const { merge } = require('webpack-merge')
@@ -8,7 +8,7 @@ const addons = (/* string | string[] */ addonsArg) => {
   let addons = [...[addonsArg]] // Normalize array of addons (flatten)
     .filter(Boolean); // If addons is undefined, filter it out
 
-  return addons.map(addonName => require(`./build-utils/addons/webpack.${addonName}.js`));
+  return addons.map(addonName => require(`./webpack/addons/webpack.${addonName}.js`));
 };
 
 module.exports = env => {
@@ -20,7 +20,7 @@ module.exports = env => {
     console.log(`${chalk.yellow('Warning! API_URL is undefined')}`);
   }
 
-  const envConfig = require(`./build-utils/webpack.${env.env}.js`);
+  const envConfig = require(`./webpack/webpack.${env.env}.js`);
   const mergedConfig = merge(
     commonConfig,
     envConfig,
