+++ web-app/webpack/rules/styles.js
@@ -1,0 +1,17 @@
+module.exports = {
+  cssLoader: {
+    loader: 'css-loader',
+    options: {
+      esModule: true,
+      modules: {
+        namedExport: true
+      }
+    }
+  },
+  sassLoader: {
+    loader: 'sass-loader',
+    options: {
+      implementation: require('sass') // Prefer `dart-sass`
+    }
+  }
+}
